Fixed a signed/unsigned problem...
CVS patchset: 6862
CVS date: 2005/07/03 21:33:42

+++ lib/stonith/st_ttylock.c
@@ -1,4 +1,4 @@
-/* $Id: st_ttylock.c,v 1.4 2005/05/13 19:16:43 gshi Exp $ */
+/* $Id: st_ttylock.c,v 1.5 2005/07/03 21:33:42 alan Exp $ */
 #include <portability.h>
 
 #include <stdio.h>
@@ -122,7 +122,7 @@ DoLock(const char * prefix, const char *lockname)
 {
 	char lf_name[256], tf_name[256], buf[LOCKSTRLEN+1];
 	int fd;
-	unsigned long pid, mypid;
+	long pid, mypid;
 	int rc;
 	struct stat sbuf;
 
@@ -150,7 +150,7 @@ DoLock(const char * prefix, const char *lockname)
 			if (sscanf(buf, "%lu", &pid) < 1) {
 				/* lockfile screwed up -> rm it and go on */
 			} else {
-				if (pid > 1 && (getpid() != pid)
+				if (pid > 1 && ((long)getpid() != pid)
 				    &&	((CL_KILL((pid_t)pid, 0) >= 0)
 					 ||	errno != ESRCH)) {
 					/* tty is locked by existing (not
